+++ Jetchat/app/src/main/java/com/example/compose/jetchat/NavActivity.kt
@@ -53,7 +53,7 @@ import java.util.Locale
 /**
  * Main activity for the app.
  */
-class NavActivity : AppCompatActivity(), RecognitionListener, TextToSpeech.OnInitListener {
+class NavActivity : AppCompatActivity(), TextToSpeech.OnInitListener {
     private val viewModel: MainViewModel by viewModels()
 
     @OptIn(ExperimentalMaterial3Api::class)
@@ -128,29 +128,29 @@ class NavActivity : AppCompatActivity(), RecognitionListener, TextToSpeech.OnIni
             checkPermission()
         }
 
+//        // --------------
+//        // Speech to Text
+//        var speech = SpeechRecognizer.createSpeechRecognizer(this)
+//        Log.i("LLM", "isRecognitionAvailable: " + SpeechRecognizer.isRecognitionAvailable(this))
+//        speech.setRecognitionListener(this)
+//        recognizerIntent = Intent(RecognizerIntent.ACTION_RECOGNIZE_SPEECH)
+//        recognizerIntent.putExtra(RecognizerIntent.EXTRA_LANGUAGE_PREFERENCE, "US-en")
+//        recognizerIntent.putExtra(RecognizerIntent.EXTRA_LANGUAGE_MODEL, RecognizerIntent.LANGUAGE_MODEL_FREE_FORM)
+//        recognizerIntent.putExtra(RecognizerIntent.EXTRA_MAX_RESULTS, 3)
+//        viewModel.setSpeechRecognizer (speech, recognizerIntent)
+//        // speech.startListening(recognizerIntent);
+//        // speech.stopListening();
 
-        // Speech to Text
-        var speech = SpeechRecognizer.createSpeechRecognizer(this)
-        Log.i("LLM", "isRecognitionAvailable: " + SpeechRecognizer.isRecognitionAvailable(this))
-        speech.setRecognitionListener(this)
-        recognizerIntent = Intent(RecognizerIntent.ACTION_RECOGNIZE_SPEECH)
-        recognizerIntent.putExtra(RecognizerIntent.EXTRA_LANGUAGE_PREFERENCE, "US-en")
-        recognizerIntent.putExtra(RecognizerIntent.EXTRA_LANGUAGE_MODEL, RecognizerIntent.LANGUAGE_MODEL_FREE_FORM)
-        recognizerIntent.putExtra(RecognizerIntent.EXTRA_MAX_RESULTS, 3)
-
-        viewModel.setSpeechRecognizer (speech)
-        //speech.startListening(recognizerIntent);
-        // speech.stopListening();
-
-        Log.d("LLM", "start TTS")
-        // TextToSpeech(Context: this, OnInitListener: this)
+        // --------------
+        // Text to Speech
         tts = TextToSpeech(this, this)
+        Log.d("LLM", "start text to speech")
         viewModel.setSpeechGenerator (tts)
         //tts.speak("Welcome to Jetchat AI", TextToSpeech.QUEUE_FLUSH, null,"")
 
     }
-    private lateinit var speech: SpeechRecognizer
-    private lateinit var recognizerIntent: Intent
+    //private lateinit var speech: SpeechRecognizer
+    //private lateinit var recognizerIntent: Intent
     val RecordAudioRequestCode = 1
     private fun checkPermission() {
         if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.M) {
@@ -177,63 +177,63 @@ class NavActivity : AppCompatActivity(), RecognitionListener, TextToSpeech.OnIni
     /**
      * Implement `RecognitionListener`
      */
-    override fun onReadyForSpeech(p0: Bundle?) {
-        Log.i("LLM", "onReadyForSpeech")
-    }
-
-    override fun onBeginningOfSpeech() {
-        Log.i("LLM", "onBeginningOfSpeech")
-    }
-
-    override fun onRmsChanged(p0: Float) {
-        //Log.d("LLM", "onRmsChanged")
-    }
-
-    override fun onBufferReceived(p0: ByteArray?) {
-        Log.d("LLM", "onBufferReceived")
-    }
-
-    override fun onEndOfSpeech() {
-        Log.d("LLM", "onEndOfSpeech")
-    }
-
-    override fun onError(p0: Int) {
-        val errorMessage: String = getErrorMessage(p0)
-        Log.e("LLM", "onError FAILED $errorMessage")
-    }
-
-    override fun onResults(p0: Bundle?) {
-        Log.i("LLM", "onResults")
-        val matches = p0!!.getStringArrayList(SpeechRecognizer.RESULTS_RECOGNITION)
-        var text = ""
-        if (matches != null) { // seems to repeat, grab one
-            for (result in matches) text = result
-        }
-        Log.i("LLM", "heard: $text")
-    }
-
-    override fun onPartialResults(p0: Bundle?) {
-        Log.d("LLM", "onPartialResults")
-    }
-
-    override fun onEvent(p0: Int, p1: Bundle?) {
-        Log.d("LLM", "onEvent")
-    }
-
-    private fun getErrorMessage(error: Int): String {
-        return when (error) {
-            SpeechRecognizer.ERROR_AUDIO -> "Audio error"
-            SpeechRecognizer.ERROR_CLIENT -> "Client error"
-            SpeechRecognizer.ERROR_INSUFFICIENT_PERMISSIONS -> "Insufficient permissions error"
-            SpeechRecognizer.ERROR_NETWORK -> "Network error"
-            SpeechRecognizer.ERROR_NETWORK_TIMEOUT -> "Network timeout error"
-            SpeechRecognizer.ERROR_NO_MATCH -> "No match error"
-            SpeechRecognizer.ERROR_RECOGNIZER_BUSY -> "Recognizer busy error"
-            SpeechRecognizer.ERROR_SERVER -> "Server error"
-            SpeechRecognizer.ERROR_SPEECH_TIMEOUT -> "Speech timeout error"
-            else -> "Unknown speech recognizer error"
-        }
-    }
+//    override fun onReadyForSpeech(p0: Bundle?) {
+//        Log.i("LLM", "onReadyForSpeech")
+//    }
+//
+//    override fun onBeginningOfSpeech() {
+//        Log.i("LLM", "onBeginningOfSpeech")
+//    }
+//
+//    override fun onRmsChanged(p0: Float) {
+//        //Log.d("LLM", "onRmsChanged")
+//    }
+//
+//    override fun onBufferReceived(p0: ByteArray?) {
+//        Log.d("LLM", "onBufferReceived")
+//    }
+//
+//    override fun onEndOfSpeech() {
+//        Log.d("LLM", "onEndOfSpeech")
+//    }
+//
+//    override fun onError(p0: Int) {
+//        val errorMessage: String = getErrorMessage(p0)
+//        Log.e("LLM", "onError FAILED $errorMessage")
+//    }
+//
+//    override fun onResults(p0: Bundle?) {
+//        Log.i("LLM", "onResults")
+//        val matches = p0!!.getStringArrayList(SpeechRecognizer.RESULTS_RECOGNITION)
+//        var text = ""
+//        if (matches != null) { // seems to repeat, grab one
+//            for (result in matches) text = result
+//        }
+//        Log.i("LLM", "heard: $text")
+//    }
+//
+//    override fun onPartialResults(p0: Bundle?) {
+//        Log.d("LLM", "onPartialResults")
+//    }
+//
+//    override fun onEvent(p0: Int, p1: Bundle?) {
+//        Log.d("LLM", "onEvent")
+//    }
+//
+//    private fun getErrorMessage(error: Int): String {
+//        return when (error) {
+//            SpeechRecognizer.ERROR_AUDIO -> "Audio error"
+//            SpeechRecognizer.ERROR_CLIENT -> "Client error"
+//            SpeechRecognizer.ERROR_INSUFFICIENT_PERMISSIONS -> "Insufficient permissions error"
+//            SpeechRecognizer.ERROR_NETWORK -> "Network error"
+//            SpeechRecognizer.ERROR_NETWORK_TIMEOUT -> "Network timeout error"
+//            SpeechRecognizer.ERROR_NO_MATCH -> "No match error"
+//            SpeechRecognizer.ERROR_RECOGNIZER_BUSY -> "Recognizer busy error"
+//            SpeechRecognizer.ERROR_SERVER -> "Server error"
+//            SpeechRecognizer.ERROR_SPEECH_TIMEOUT -> "Speech timeout error"
+//            else -> "Unknown speech recognizer error"
+//        }
+//    }
 
     /**
      * Implement speech output
